refactor(engine): use engine-core public APIs in engine-dom
